requirements.txt
aiofiles==23.1.0
aiosqlite==0.18.0
anyio==3.6.2
APScheduler==3.10.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.90.0
fastapi-amis-admin==0.4.3
fastapi-scheduler==0.0.11
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.7
pyasn1==0.4.8
pydantic==1.10.5
pyjsbn-rsa==0.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.41
sqlalchemy-database==0.1.0
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
starlette==0.23.0
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
alembic
